+++ NEWS.md
@@ -1,7 +1,9 @@
 # branch no-phangorn (9002)
 
 - Internal implementation of path and SPR distances, removing dependency
-  on phangorn (and thus R 4.1)
+  on phangorn (and thus R 4.1).
+  
+- `PathVector()` now treats trees with a root node as rooted.
 
 # TreeDist 2.3.0.9001